Feature/auth (#9)
* 疎通確認用スクリプトの更新

* redis周りの更新

* usecaseの修正

* env周りの修正

* DB周りの修正

* redisの操作関連コードを作成

* 不要な行を削除

* 改行の追加

* TokenServiceの引数を緩和

* 引数を緩和

* 認証まわりの追加実装

* 認証のエンドポイントテストを追加

* 追跡を削除

* 改行を追加

* ログアウト機能の追加

+++ containers/application/backend/src/usecase/user/DeleteUserUseCase.ts
@@ -2,16 +2,22 @@ import type { UserRepository } from '../../domain/user/repository/UserRepository
 import type { DeleteUserForm } from '../../domain/user/form/UserForm.ts';
 import { getUnixTimeMs } from '../../utils/unixtime.ts';
 import UserId from '../../domain/user/vo/UserId.ts';
+import Password from '../../domain/user/vo/Password.ts';
 
 export class DeleteUserUseCase {
   constructor(private repo: UserRepository) {}
 
   async execute(id: UserId, form: DeleteUserForm): Promise<void> {
-    // TODO: user authnication
     const user = await this.repo.findById(id.get())
     if (!user) {
       throw new Error("user not found")
     }
+
+    if (!Password.compare(form.password, user.password)) {
+      console.warn("LoginUseCase: Password compare is failed.")
+      throw new Error("Invalid email or password.")
+    }
+
     const now = getUnixTimeMs();
     const deletedEmail = `${form.email}_${now}`;
     await this.repo.delete(user.id, deletedEmail, now)
